+++ packages/validator-ajv8/test/createPrecompiledValidator.test.ts
@@ -30,7 +30,7 @@ describe('createPrecompiledValidator()', () => {
       expect(custom).toBeInstanceOf(AJV8PrecompiledValidator);
     });
     it('precompiledValidator was constructed with validateFns and rootSchema', () => {
-      expect(AJV8PrecompiledValidator).toHaveBeenCalledWith(validateFns, rootSchema, undefined);
+      expect(AJV8PrecompiledValidator).toHaveBeenCalledWith(validateFns, rootSchema, undefined, undefined);
     });
   });
   describe('passing validatorFns, rootSchema and localizer to createPrecompiledValidator', () => {
@@ -45,7 +45,20 @@ describe('createPrecompiledValidator()', () => {
       expect(custom).toBeInstanceOf(AJV8PrecompiledValidator);
     });
     it('defaultValidator was constructed with validateFns, rootSchema and the localizer', () => {
-      expect(AJV8PrecompiledValidator).toHaveBeenCalledWith(validateFns, rootSchema, localizer);
+      expect(AJV8PrecompiledValidator).toHaveBeenCalledWith(validateFns, rootSchema, localizer, undefined);
+    });
+  });
+  describe('passing suppressDuplicateFiltering to createPrecompiledValidator', () => {
+    let custom: any;
+    beforeAll(() => {
+      mockedValidator.mockClear();
+      custom = createPrecompiledValidator<TestType>(validateFns, rootSchema, undefined, 'all');
+    });
+    it('precompiled validator was created', () => {
+      expect(custom).toBeInstanceOf(AJV8PrecompiledValidator);
+    });
+    it('precompiledValidator was constructed with validateFns, rootSchema, undefined, and suppressDuplicateFiltering', () => {
+      expect(AJV8PrecompiledValidator).toHaveBeenCalledWith(validateFns, rootSchema, undefined, 'all');
     });
   });
 });